dom/tests/mochitest/bugs/test_bug465263.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=465263
-->
<head>
  <title>Test for Bug 465263</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=465263">Mozilla Bug 465263</a>
<p id="display">
  <iframe src="http://mochi.test:8888"></iframe>
  <iframe src="http://mochi.test:8888"></iframe>
  <iframe src="http://mochi.test:8888"></iframe>
  <iframe src="http://mochi.test:8888"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 465263 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  window.frames[0].location.hash = '';
  is(window.frames[0].location.href, "http://mochi.test:8888/#",
     "Should have '#' in href now");
  window.frames[1].location.hash = '#';
  is(window.frames[1].location.href, "http://mochi.test:8888/#",
     "Should have only one '#' in href");
  window.frames[2].location.hash = 'foo';
  is(window.frames[2].location.href, "http://mochi.test:8888/#foo",
     "Should have '#foo' in href");
  window.frames[3].location.hash = '#foo';
  is(window.frames[3].location.href, "http://mochi.test:8888/#foo",
     "Should have only one '#' in href here too");
  SimpleTest.finish();
});




</script>
</pre>
</body>
</html>
